tests

+++ cryptoapp/views.py
@@ -1,11 +1,25 @@
 from django.shortcuts import render
-import hashlib, random, string, uuid
 from django.core.mail import send_mail
-from cryptoapp.models import Profile, Address, Message, Order, Validation_token
 from django.contrib.auth.models import User
-from rest_framework.views import APIView
-from knox.auth import TokenAuthentication
+from django.contrib.auth import authenticate
+from cryptoapp.models import Profile, Address, Message, Order, Validation_token
 from rest_framework.permissions import IsAuthenticated
+from rest_framework.response import Response
+from rest_framework.views import APIView
+from rest_framework import status
+from knox.models import AuthToken
+from knox.auth import TokenAuthentication
+import hashlib, random, string, uuid, selenium, json, datetime
+
+
+class placeOrderForm(APIView):
+
+    def post(self, request, format=None):
+        try:
+            return
+        except:
+            return
+            #take selenium screenshot of url yea.
 
 class changeData(APIView):
     authentication_classes = (TokenAuthentication,)
@@ -89,12 +103,11 @@ class recoverPassword(APIView):
 
     def post(self, request, format=None):
         try:
-            profile = User.objects.get(email=request.data['email'])
+            user = User.objects.get(email=request.data['email'])
             salt = hashlib.sha1(str(random.random()).encode('utf-8')).hexdigest()[:15]
-            user = User.objects.get(email=email)
             user.set_password(str(salt))
             message = 'this is your new password, please login and change it %s' % str(salt)
-            send_mail('btc buyer recover password', message, 'defaultemail@email.com', [str(email)], fail_silently=False)
+            #send_mail('btc buyer recover password', message, 'defaultemail@email.com', [str(email)], fail_silently=False)
             return Response({'message': 'email sent'}, status=status.HTTP_200_OK)
         except Exception as e:
             print('recover password post %s' % e)
@@ -107,10 +120,12 @@ class deleteUser(APIView):
 
     def post(self, request, format=None):
         try:
-            user = request._auth.user
-            user.auth_token_set.all().delete()
-            user.delete()
-            return Response({'user': user.username, 'message': 'user deleted'}, status=status.HTTP_200_OK)
+            user = authenticate(username=request.data['email'], password=request.data['password'])
+            print('authed?')
+            if user is not None and request._auth.user == user:
+                user.auth_token_set.all().delete()
+                user.delete()
+                return Response({'user': user.username, 'message': 'user deleted'}, status=status.HTTP_200_OK)
         except Exception as e:
             print('delete user post %s' % e)
             return Response({'message':'user deletion error'}, status=status.HTTP_400_BAD_REQUEST)
@@ -130,7 +145,9 @@ class validateEmail(APIView):
 
     def post(self, request, format=None):
         try:
-            return_data = Misc.send_validation_email(self, request.data['email'])
+            user = User.objects.get(email=request.data['email'])
+            profile = Profile.objects.get(user=user)
+            return_data = Misc.send_validation_email(self, request.data['email'], profile.uuid)
             return Response({'message': 'email sent'}, status=status.HTTP_200_OK)
         except Exception as e:
             print('validate email post %s' % e)
@@ -140,11 +157,11 @@ class validateEmail(APIView):
     def get(self, request, *args, format=None):
         try:
             token = Validation_token.objects.get(token=args[0])
-            now = timezone.now()
-            token_expiry = token.dateCreated + datetime.timedelta(days=1)
-            if now < token_expiry:
+            now = datetime.datetime.utcnow() + datetime.timedelta(seconds=1)
+            if datetime.datetime.utcnow() < token.expires.replace(tzinfo=None):
                 profile = Profile.objects.get(uuid=token.user_uuid)
                 profile.email_validated = True
+                token.delete()
                 profile.save()
                 return Response({'message': 'email validation success'}, status=status.HTTP_200_OK)
             else:
@@ -156,36 +173,47 @@ class validateEmail(APIView):
 
 class Misc(APIView):
 
-    def send_validation_email(self, email):
+    def create_user(self, request):
+        try:
+           uu = uuid.uuid4()
+           if User.objects.filter(email=request.data['email']).exists() == False:
+               try:
+                   first_name = request.data['firstName']
+               except:
+                   first_name = ''
+               try:
+                   last_name = request.data['lastName']
+               except:
+                   last_name = ''
+               user = User.objects.create_user(request.data['email'], email=request.data['email'],  password=request.data['password'], last_name=last_name, first_name=first_name)
+           else:
+               return Response(status=status.HTTP_400_BAD_REQUEST)
+           user.save()
+           try:
+               phone_number = request.data['phone_number']
+           except:
+               phone_number = ''
+           profile = Profile.objects.create(user=user, phone_number=phone_number)
+           Misc.send_validation_email(self, request.data['email'], profile.uuid)
+           token = AuthToken.objects.create(user)
+           return_data = {'user': user.email, 'message': 'email sent', 'token': token}
+           return return_data
+        except Exception as e:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+
+
+    def send_validation_email(self, email, uuid):
         try:
             salt = hashlib.sha1(str(random.random()).encode('utf-8')).hexdigest()[:30]
             user = User.objects.get(email=email)
-            token = Validation_token.objects.create(user_uuid=user.uuid, token=salt)
+            if Validation_token.objects.filter(user_uuid=uuid).exists() == True:
+                Validation_token.objects.filter(user_uuid=uuid).delete()
+            expires = datetime.datetime.utcnow() + datetime.timedelta(days=1)
+            token = Validation_token.objects.create(user_uuid=uuid, token=salt, expires=expires)
             url = 'theurl.com/api/v1/validate/%s' % salt
             token.save()
-            send_mail('btc buyer email validation', url, 'defaultemail@email.com', [str(email)], fail_silently=False)
+            #send_mail('btc buyer email validation', url, 'defaultemail@email.com', [str(email)], fail_silently=False)
             return True
         except Exception as e:
             print('misc send_validation_email %s' % e)
             return False
-
-    def create_user(self, request):
-        try:
-           uu = uuid.uuid4()
-           if User.objects.filter(email=request.data['email']).exists():
-               user = User.objects.create_user(request.data['email'], email=request.data['email'],  password=request.data['password'])
-           else:
-               return Response({'message':"that email is taken"}, status=status.HTTP_400_BAD_REQUEST)
-           user.save()
-           self.send_validation_email(request.data['email'])
-           if request.data['phone_number'] is not None:
-               phone_number = request.data['phone_number']
-           else:
-               phone_number = ''
-           profile = Profile.objects.create(user=user, phone_number=phone_number)
-           token = AuthToken.objects.create(user)
-           return_data = {'user': str(request.data['name']), 'message': 'email sent'}
-           return return_data
-        except Exception as e:
-            print('misc create_user %s' % e)
-            return Response({'message':"something went terribly wrong"}, status=status.HTTP_400_BAD_REQUEST)
